Adds firestore integration, and proper typescript
TypeScript integration was properly added using the vue
cli.

Firestore was added, kind of.
It doesn't automatically update the DOM when data
is added or deleted. This is due to the fire store method
causing typing issues so the create method was used instead.
I will need to investigate further to figure out what is wrong.

BaseInputText was deleated was it was overly complicated and
unnesscary.

Version is now 0.2.0