Fixes firestore usage.
Now uses the firstore method instead of the create
method to get data from FireStore.
It turns out that the .sortby() method returns a
Query object, which does not mesh with the firestore method.
Sorting will have to be figured out at some other time.
For now I'm happy that it updates the web page when items are created/destroyed.

Now version 0.3.0