Adds firebase authentication
Followed this (medium article)[https://medium.com/dailyjs/authenticating-a-vue-js-application-with-firebase-ui-8870a3a5cff8]
The list of chores is only viewable upon logging in now.
Additionally the user must be a part of the list that I allow to read/write data in order for the app to populate.
Version is now 0.6.0